fix: partition number range pattern tests
