+++ packages/ecma402-abstract/tests/FormatApproximately.test.ts
@@ -1,29 +1,6 @@
 import {FormatApproximately} from '../NumberFormat/FormatApproximately'
-import {NumberFormatInternal, NumberFormatPart} from '../types/number'
-
-const internalSlotMap = new WeakMap<Intl.NumberFormat, NumberFormatInternal>()
-
-function getInternalSlots(x: Intl.NumberFormat): NumberFormatInternal {
-  let internalSlots = internalSlotMap.get(x)
-  if (!internalSlots) {
-    internalSlots = Object.create(null) as NumberFormatInternal
-    internalSlotMap.set(x, internalSlots)
-  }
-  return {
-    ...internalSlots,
-    numberingSystem: 'latn',
-    dataLocaleData: {
-      numbers: {
-        symbols: {
-          // @ts-ignore
-          latn: {
-            approximatelySign: '~',
-          },
-        },
-      },
-    },
-  }
-}
+import {NumberFormatPart} from '../types/number'
+import {getInternalSlots} from './utils'
 
 describe('FormatApproximately', () => {
   const numberFormat: Intl.NumberFormat = new Intl.NumberFormat('it')
